Make test workflow driver actually async (#68)
This makes the test workflow driver actually pause when commands are awaited. This ultimately is more complicated than it was before, but is certainly not a waste as this is pretty much what's going to have to happen in a real Rust workflow.

We could also use this in the higher level tests with some additional changes, which would obviate the need to manually issue completions.