Add integ test. Add timer backoff logic to WfContext. WfContext non-mut.
The local activity function which does timer-based backoff in the SDK
cannot take WfContext mutably without making huge issues for users
who want to start it and then do something else concurrently. WfContext
did everything with &mut because there was no need until now for that
to be different. Now there is. The conversion is very simple.