Bump dashmap version in Cargo.toml
We're using APIs that are only available in v5.5 (e.g. .insert() is
used in core/src/worker/activities.rs). In practice, when compiling
sdk-core alone, v5.5 was already what was being picked up and there
was no problem. However, when sdk-core is compiled as a crate in the
sdk-typescript core-bridge project, v5.4 is getting picked up, with
the result that sdk-core fails to build in that context.