Expose the `WorkerClient` trait (#924)
* Expose the `WorkerClient` trait

# Motivation
The `WorkerClient` trait encapsulates all the methods that the worker needs to make the underlying API calls. Currently, it is only exposed to the crate. However, it is useful to expose this publicly to allow other language clients to provide their native `WorkerClient` implementation e.g. by backing it with the language specific gRPC library.

# Modifications

This PR makes the `WorkerClient` trait and other types public.

# Result

Other language SDKs can now substitute in their gRPC client.

* Format

* Skip cloud test on forks properly

---------

Co-authored-by: Spencer Judge <spencer@temporal.io>