Handle NexusOperationCancelRequest{Completed, Failed}
These events are written by the caller server on receiving a response
from the nexus cancel handler. The only cancellation type that is
sensitive to them is WaitRequested: when that is in effect, these
events cause an activation job to be emitted that resolves the nexus
operation future.

The fsm macro is modified to allow the same union type to be
constructed twice, since Started -> Started|Cancelled is now caused by
two different inputs.